Fix charsets
I didn't expect I was going to have to actually spend time editing every
single HTML file, but the meta tags (and the head sections, and
sometimes the HTML tags) were a mess pretty universally. If you've got
one, a !DOCTYPE must be first. And it's about the only tag that should
be in caps prescriptively. (For all others it's become convention not
to, but standards-wise it's up to you.)

After !DOCTYPE, you should have an html tag block. Within that should be
a head block, and the very first thing in your head block should be a
meta tag to tell the browser what character set you're using! It'll do
some heuristics and fall back on the server up until that point, but
like setting the charset is the first thing you do before even the title
(since the title might depend on the charset!)

Beyond that, go nuts, especially in quirks mode. Which of course all of
these pages are in quirks mode.

That should be the last of the HTML errors. The image assets aren't
there yet and the site still picks up stuff from outside its local tree.
I'll address those tomorrow. This is taking a lot more effort than I
realized because I'm trying to fix some of the biggest problems with the
site that aren't just old content while I'm at it.