Tell Github not to bother running Jekyll
We're using Github Pages as a staging area to be able to see what was
done to the tree kinda fast. But Github Pages takes the time to run
Jekyll on the site, a template pre-processor that's really handy if
you're using it. But we aren't. This .nojekyll file tells Github not to
bother, which should speed up deployment of changes to Github Pages.