Introducing staging lastline comment
I am going through these files one by one and adding their images to the
archive. This is kind of phase 2 of adding Agin's stuff to git. Stage 3
will be a little more complex, and I haven't decided exactly how I
intend to tackle it yet (dealing with links…)

The last thing I'm doing as I finish these is files is add a comment to
the bottom which is designed to be machine-readable indicating the
file's finished stage 2. That comment looks like this:

<!-- ##2## -->

I'll be adding that to each file I add images for as I do that going
forward.