Add images for several html files
Basically, every webpage in the tree linked directly from index.html has
now had its images added. (I hope.)

This is quite a mess already, I think we might be forced to do some
cleanups/reorganizations just because this is already a _MESS_. And it
is going to get worse.