Add .editorconfig and .gitattributes
This resolves #4. The .editorconfig demands that text files be
terminated in now-everything-but-Windows style, which even Windows 10
now supports. This means overriding a stupid behavior of git to bend
over backwards to try and give Windows users CRLF for … at this point
old versions of Windows and Visual Studio, and not much else.