Mark orphaned HTML files
Used the same kind of ##-delimited comment I'm using for the staged
updates to mark the orphan files. For the moment marking them to be
skipped in that process for now.

Fixed a couple of minor problems I found along the way (character set
and failing to close body/html tags…) I definitely didn't bother to
actually clean up the HTML on these or anything.