fix index.html's DTD line
Yes, DTD lines like this are obsolete. I still want it to be *correct*
so that I can throw a validator that still understands these old files
at it and maybe have it show me some of the broken-ass tag soup in an
automated way and let me fix it.