Stage 2 for most of the HTML files
A handful yet remain. These files are such a haphazard mess I just can't
leave them like this. Part of what I've been doing is determining what's
part of Agin's Inn, what is lost media, and what's duplicated, old,
unwanted crap left over from trying to manage this with ftp. But sifting
through this and getting actual content out of it … not as a website but
as a file tree … no, it's a disaster!

Better organization is needed. We can 301 the HTML files easily enough.
And the assets can be managed in a way that makes sense. But it doesn't
make sense to do that without cleaning up the HTML too as long as we can
do it without changing the look of the site appreciably.

We'll figure it out.