Rename .main-page to .index-page
Also give myself a paper bag over the head for failing to ../ my URL in
the css file I put into a subdir. See squidgemonster? Everyone's code is
shit, it's just the nature of the beast. One must simply fix it, declare
themselves to be officially an idiot, and then go on to make some OTHER
horrifyingly dumb mistake, since you're not likely to repeat this one.
At least not for the rest of the day. 😁