Minor tag cleanup (lowercased), forgot css
Forgot the CSS to go along with renaming the index page's big div.

Also, I lowercased the tags (not their arguments because I couldn't do
that programmatically) throughout. Might help with comparisons to clean
up some orphan files.

Last change I made was putting all the keyword metas on a single line in
each file. This made a grep easy to see that the site basically has one
set of keywords used on almost every file. Which means I can move it to
a template, because that's probably happening soonish.