Add .gitignore with jekyll bits
So … we're gonna use Jekyll. This doesn't enable it on the site yet
(that's next commit), but this disables the kind of obnoxious droppings
that get into git's way. I'm going to need to set up something to be
fairly cookie cutter for installing the github-pages Jekyll toolchain
for Windows at least. Maybe one of the other devs can provide some help
there?