Starting to build some CSS
I've made the index page use the same <div> wrapper I'm going to use to
replace the <center> tag at the top level on all the pages, but gave it
an id so that it can have a custom background style (and so that it will
pick up the black link colors in a bit less half-assed manner.)

We can't depend on the user having the standard black on white with
blue/purple/red color scheme. It uses those, along wwith table layouts,
everything in <center>, a bunch of <hr> everywhere … classic GeoCities
look. We'll recreate it.

This doesn't use Jekyll yet because the index page is a little special.
I'll do a few more top-level pages and get a feel for it, then I'll come
back and modify things to work with the index.

I think I'm going to do my sizing on a lot of things in em because I'm
going to use some cleverness to handle hidpi later on.