Make CSS reload on site build
This will cause the cached CSS to invalidate every time the page is
rebuilt. This … is good enough. Once this site is converted to html5, it
won't be changing but seldomly. In the meantime, yes, I want it to
invalidate the cache every time the site is regenerated!

This trick may not be ideal for YeOldeInn … but there the CSS files are
not likely to be large enough that it's a problem.