Add .htaccess to site
This is basically the .htaccess live on the server right now, but not
quite the one that was live yesterday. Basically the original used
index.php for the page, and that was a frame that sort of hid what page
you were on. It was annoying, it could've been used to have a "Back to
YeOldeInn" link or something, but it didn't. No, it just made links for
comparison purposes harder, so I got rid of it.

It seems to be necessary to specify that .php is not wanted anymore, so
I'm 301ing it.

The ErrorDocuments don't work and I'm actually removing them in my next
commit, but I want them here for the space of one commit.