Wrap content in a <main>, use CSS vars
Changed new.html's <main> to a <section> (no CSS changes needed).
Wrapped {{ content }} in a <main> because it allows me to specify its
width and centeredness.

While I was in there, converted colors to CSS vars. Makes it trivial to
go in and dark mode things later on without affecting the whole
GeoCities vibe. That's later.