Start work on quests.html
Starting to convert quests.html to HTML5. That caused me to make some
changes to how new.html handles some things … there's going to be a bit
of going back and changing stuff until we get the hang of this, but I
think I'm starting to get the idea of what it should look like now.

It's easy to start putting paragraph text into paragraphs, but a major
thing I haven't decided how to deal with yet is the whole "yeah that
link has been dead for a decade" problem. I fixed it by making a nolink
span that looks like a link. Use it on something that used to be a link
and then do something with the title attribute to explain what needs
explaining. Did that for Dewayne's defunct innkeep@ email. This was a
minor content edit because I turned it into a nolink (which changed its
color) and removed the instructions for how to make the email address
work because … no need for that now.

There's a lot more to do with this since the crazy big ugly quests table
doesn't fit in the 1008 width given a 16px base font. I could change
that since most people use a 12px base font, but I _hate_ pages that do
that, so I'm not going to. No, I can see what's causing the page to be
too long and I'll be addressing that when I get there.

Figuring out how to CSS the ugly-ass GeoCities era table layout was fun.