NO MORE .htm FILES!
Every .htm file has been renamed to .html now. No more need to
constantly make commands that look for both extensions. This doesn't
move the needle even a tiny bit in terms of making stuff more HTML5, but
it makes it easier for me to make changes in the future without missing
anything.

…assuming I didn't miss anything.