So tiles.html is HTML5 (stage 3), ish.
Cheated and left the stupid table in the middle (I did set its margins
to auto, though I did it lazily inline), but this isn't the right way to
do it. No, the right way to do it would be to specify a flex that causes
these to expand up to 3 columns with a maximum width that seems
appropriate to the flow of the page. No bigger, but allowed to collapse
to be 2 or even 1 column as the page shrinks.

It'd be really freakin' cool to do that.

And I don't reliably know how to do it, which means this close to
bedtime is NOT the time for me to once again attempt to understand how
to use flexbox properly. So I'm leaving it for now.