Interim totally not right or done minis.html
Committing this work in progress because it's pretty cool. The page
resizes! Both Chrome and Firefox have a way to do this. It's
ctrl-shift-M on Windows/Linux and ⌘⇧M on Mac. A simple and cool thing
happens when you cross the 600px wide barrier. The page reformats. Oh,
and the header graphic shrinks to fit. I mean, I'm impressed by this
because I'm keeping the whole GeoCities vibe, but … modern features
under the hood.