Moved minis.html CSS and fun w/ HiDPI!
I moved the CSS in minis.html to agins.scss—I think I'll eventually be
breaking down this "for one page" stuff into its own file, let's see how
that shakes out over time, but for now putting it in the stylesheet is
the right place so it can be refactored if needed.

Did something fun! Most of the pages on Agin's Inn have these GIF header
images that look quite pixelated and blocky today. Some of the effect is
lost when your brower upscales them using at least bilinear if not
bicubic interpolation. Let's encourage them to try and keep all that
pixelated glory intact… 🤣 I created integer 2x and 4x nearest neighbor
resizes. The original is smaller as a GIF, but the scales work better as
PNG, especially after a pngcrush, so that's what we're going with.

Of course, that's gonna get really unwieldy and we've already got images
all over the freakin' place, so I've now begun the process of organizing
the image assets in this site. One small step for nerd, one giant leap
for nerdkind, that sort of thing. The assets directory will eventually
encompass all of the non-HTML page elements, in a nice organized
fashion.

Moving this file meant a lot of other files (sub-pages of the minis
gallery) needed to be edited, so I made them use the HiDPI images early
for just the hader.