Migrating to new quest table (Unfinished)
This has been over a week in the making and I'm not finished with it
yet. Why commit it now? Because I almost fucked up. I deleted
quests.html. Fortunately it was open in my editor and I could just save
a new copy. The best backup in the world is to put your code on the
Internet.

So, there are 21 quests missing here. That's … less than half, but it's
still 40% of the quests needing to be converted to "markdown" format.

I say "markdown" because … none of these markdown files in the _quests
collection actually contain "markdown" content. The files are entirely
yaml front matter. Sure, a fair bit of it is written in Markdown, but
ideally the reqs would be reworked so they were done similarly to the
filenames. I just can't be arsed to do it—it's taken me long enough to
do them as they are right now!

Yes, I've been sick, which has had my stuffed up head feeling kind of
fuddled. Yes, I've had other things happening. But still, I need this
to get done.

Jekyll, the software that is builting the website on Github, has a
notion of "collections". If you know anything about Jekyll, the _posts
directory is a collection that's hardcoded to exist and defaults to
output: True. Others you have to define and output: False is the default
(which means they're not rendered into _site with the rest of your
output…)

I created _quests as an output: false collection. I can still access all
the front matter andd page content from another page (which is how you
get a blog's listing page made from individual posts), but on this site
the individual quest packs don't get their own pages. That's why the
content is blank. If anything the content should be the description of
the pack I guess, but I decided to make it part of the front matter too,
at least for the time being.

It's easy to change, but I doubt I will until I start doing a quests
page for YOI. I'm trying to change as little as possible on Agin's Inn
since it's supposed to be an archive. I'm just having to change a lot
more than I bargained for to make the content work everywhere.

Oh yeah, I buried the lede on that one pretty deep. If you're on
Windows, press ctrl-shift-on the github.io agins-archive page. Set the
page width to 900, then go to New Quests. Change the width to 899. Isn't
that freakin' cool? The entire table (the one I've been working on
anyway) changes from a really wide horizontal table to a list of tables
each containing one quest, vertically.

And yes, it sizes down! It works all the way down to 400px! Smaller than
400, it breaks. *sigh* Fixing that is low priority, but I can fix it. I
wouldn't bother except I think I know why it doesn't work at 400, and
I've made other things work all the way down to 320. If your device
doesn't have 320px width … um.