Big massive art commit
The idea I was going for here was that rather than adding all the art
assets to the tree as the mess they are, I would have them separate and
put them in place as I encounter them. So, when I encounter an image
that's not yet where it ought to be, I'll decide where it will live.
Upon doing so, I'll move it there from the not-yet-converted directory
and make all the changes to the requisite HTML files.

When I'm done, I'll have only orphans left in not-yet-converted/ and can
decide what to do with them.

When it was just me, there was no point in added them to the tree and
making a mess out of that. But now, if I'm gonna have some help, it is
worth adding them. So here they are.